Add conjecture_sequence function
Create a function named conjecture_sequence in order to place the
sequence algorithm (3n+1: if n is odd. n/2: if n is even) in a special
function outside of 'main'. Call 'conjecture_sequence' in the 'main'
function replacing the multiple lines of code that is now included in
the 'conjecture_sequence' function.